requirements.txt
hyperliquid-python-sdk>=1.0.0
web3>=6.0.0
eth-account>=0.10.0
structlog>=23.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
asyncio
statistics 